Changed: Updated unrar

+++ jni/unrar/suballoc.hpp
@@ -62,7 +62,7 @@ class SubAllocator
     void* AllocUnitsRare(int indx);
     inline RAR_MEM_BLK* MBPtr(RAR_MEM_BLK *BasePtr,int Items);
 
-    size_t SubAllocatorSize;
+    unsigned long SubAllocatorSize;
     byte Indx2Units[N_INDEXES], Units2Indx[128], GlueCount;
     byte *HeapStart,*LoUnit, *HiUnit;
     struct RAR_NODE FreeList[N_INDEXES];
@@ -78,7 +78,7 @@ class SubAllocator
     inline void* ExpandUnits(void* ptr,int OldNU);
     inline void* ShrinkUnits(void* ptr,int OldNU,int NewNU);
     inline void  FreeUnits(void* ptr,int OldNU);
-    size_t GetAllocatedMemory() {return(SubAllocatorSize);};
+    unsigned long GetAllocatedMemory() {return(SubAllocatorSize);};
 
     byte *pText, *UnitsStart,*HeapEnd,*FakeUnitsStart;
 };
